ajout token et controle token

+++ back/controllers/sauces.js
@@ -41,15 +41,29 @@ exports.getOneSauce =  (req, res, next) => {
 };
 
 exports.deleteOneSauce = (req, res, next) => {
-    Sauce.deleteOne({
-            _id: req.params.id
-        })
-        .then(() => res.status(200).json({
-            message: 'Sauce supprimée !'
-        }))
-        .catch(error => res.status(400).json({
-            error
-        }));
+    Sauce.findOne({ _id: req.params.id}).then(
+        (sauce)=>{
+            if(!sauce){
+                res.status(404).json({
+                    error: new Error('Sauce non trouvée !')
+                });
+            }
+            if(sauce.userId !== req.auth.userId){
+                return res.status(401).json({
+                    error : new Error ('Requête non autorisée !')
+                });
+            }
+            Sauce.deleteOne({
+                _id: req.params.id
+            })
+            .then(() => res.status(200).json({
+                message: 'Sauce supprimée !'
+            }))
+            .catch(error => res.status(400).json({
+                error
+            }));
+        }
+    )
 };
 
 exports.getAllSauces = (req, res, next) => {
